same naming as in collstats
